requirements.txt
google-genai
google-generativeai
pydantic
python-dotenv
pillow
tenacity
playwright
